Change test directory structure to increase clarity
* `tests/app` dir shadowed flask 'builtin' name app. Changed to application to avoid name clashes.
* Moved `tests/app/helpers.py` -> `tests/helpers.py`.
  * These helpers are more visible in the root of the test dir
  * Promote reuse and expansion.
* Split `helpers.py` into `helpers.py` and `bases.py`
  * This code was becoming enormous an multi purpose.
* Moved db setup code from `tests.app.__init__` to `conftest.py`
  * This code is called at the start ove every session. It shouldn't be nested.
