+++ tests/conftest.py
@@ -1,12 +1,17 @@
 from __future__ import absolute_import
 
+import os
 from datetime import datetime
 from itertools import product
-from six import iteritems, iterkeys
-import pytest
-import mock
 
-from .app import setup, teardown
+import mock
+import pytest
+from alembic.command import upgrade
+from alembic.config import Config
+from flask.ext.migrate import Migrate, MigrateCommand
+from flask.ext.script import Manager
+from six import iteritems, iterkeys
+from sqlalchemy import inspect
 
 from app import create_app
 from app.models import (
@@ -16,7 +21,32 @@ from app.models import (
 
 @pytest.fixture(autouse=True, scope='session')
 def db_migration(request):
-    setup()
+    print("Doing db setup")
+    app = create_app('test')
+    Migrate(app, db)
+    Manager(db, MigrateCommand)
+    ALEMBIC_CONFIG = \
+        os.path.join(os.path.dirname(__file__),
+                     '../migrations/alembic.ini')
+    config = Config(ALEMBIC_CONFIG)
+    config.set_main_option(
+        "script_location",
+        "migrations")
+    with app.app_context():
+        upgrade(config, 'head')
+    print("Done db setup")
+
+    def teardown():
+        app = create_app('test')
+        with app.app_context():
+            db.session.remove()
+            db.engine.execute("drop sequence suppliers_supplier_id_seq cascade")
+            db.drop_all()
+            db.engine.execute("drop table alembic_version")
+            insp = inspect(db.engine)
+            for enum in insp.get_enums():
+                db.Enum(name=enum['name']).drop(db.engine)
+            db.get_engine(app).dispose()
     request.addfinalizer(teardown)
